+++ app.js
@@ -2,16 +2,12 @@ const express = require('express');
 const path = require('path');
 const mongoose = require('mongoose');
 const ejsMate = require('ejs-mate');
-const { campgroundSchema, reviewSchema } = require('./schemas.js')
-const catchAsync = require('./utils/catchAsync');
 const ExpressError = require('./utils/ExpressError');
 const globals = require('./globals');
 const methodOverride = require('method-override');
-const Campground = require('./models/campground');
-const Review = require("./models/review");
-const review = require('./models/review');
 
 const campgrounds = require("./routes/campgrounds")
+const reviews = require("./routes/reviews")
 
 const dbConnection = globals.dbConnection;
 
@@ -45,38 +41,12 @@ const validateCampground = (req, res, next) => {
     }
 }
 
-const validateReview = (req, res, next) => {
-    const { error } = reviewSchema.validate(req.body);
-    if (error) {
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400)
-    } else {
-        next();
-    }
-}
-
 app.use("/campgrounds", campgrounds)
+app.use("/campgrounds/:id/reviews", reviews)
 
 app.get('/', (req, res) => {
     res.render('home')
 })
-
-
-app.post("/campgrounds/:id/reviews", validateReview, catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
-    const review = new Review(req.body.review);
-    campground.reviews.push(review);
-    await review.save();
-    await campground.save();
-    res.redirect(`/campgrounds/${campground._id}`)
-}))
-
-app.delete("/campgrounds/:id/reviews/:reviewId", catchAsync(async (req, res) => {
-    const { id, reviewId } = req.params;
-    await Campground.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
-    await Review.findByIdAndDelete(req.params.reviewId);
-    res.redirect(`/campgrounds/${id}`);
-}))
 
 app.all("*", (req, res, next) => {
     next(new ExpressError('Page Not Found', 404))
