+++ routes/campgrounds.js
@@ -5,7 +5,8 @@ const catchAsync = require('../utils/catchAsync');
 const ExpressError = require('../utils/ExpressError');
 const Campground = require('../models/campground');
 
-const { campgroundSchema, reviewSchema } = require('../schemas.js')
+
+const { campgroundSchema } = require('../schemas.js')
 
 
 const validateCampground = (req, res, next) => {
